+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Added
 *   Webhook for track\_updated [#197](https://github.com/macder/medusa-fulfillment-shippo/issues/197)
 *   Public methods for retrieving return labels [#198](https://github.com/macder/medusa-fulfillment-shippo/issues/198)
 *   Experimental method to retrieve transaction for return label [#202](https://github.com/macder/medusa-fulfillment-shippo/issues/202)
+*   Boolean method for checking if transaction is a return label [#208](https://github.com/macder/medusa-fulfillment-shippo/issues/208)
+*   ShippoOrderService [#209](https://github.com/macder/medusa-fulfillment-shippo/issues/209)
 
 Removed